backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
aiomysql==0.2.0
redis[hiredis]==6.2.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
httpx==0.28.1
python-multipart==0.0.20
anthropic==0.52.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-dysmsapi==2.1.2
tencentcloud-sdk-python-common==3.0.1325
tencentcloud-sdk-python-sms==3.0.1207
openai>=1.40.0
google-genai>=1.0.0
pyasn1>=0.4.8,<0.5.0
pyasn1-modules>=0.2.1,<0.3.0
pydantic==2.11.1
pydantic-settings==2.9.1
chromadb==0.5.7
PyYAML==6.0.2
faster-whisper>=1.0.0
alembic==1.18.3
